Week-xx-Vibe-Coding-101/hussam_app_example/requirements.txt
streamlit>=1.28.0
diffusers>=0.21.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.5.0
accelerate>=0.24.0
safetensors>=0.4.0
